requirements.txt
pymupdf
python-dotenv
langchain
langchain_groq
langchain_core